+++ src/index.js
@@ -228,6 +228,25 @@ class KeyringController extends EventEmitter {
         }
     }
 
+    importWallet(_privateKey) {
+        try {
+            
+            if (_privateKey.startsWith('0x')) {
+                _privateKey = _privateKey.slice(2)
+            }
+            const privateKey = Buffer.from(_privateKey, 'hex')
+
+            if (!ethUtil.isValidPrivate(privateKey))
+                throw "Enter a valid private key"
+
+            const address = ethUtil.bufferToHex(ethUtil.privateToAddress(privateKey))
+            this.importedWallets.push(address);
+            return address
+        } catch (e) {
+            return Promise.reject(e)
+        }
+    }
+
 
     //
     // PRIVATE METHODS